moved tests from /routes to /tests
